(dsa) improved handling of memcpy
By default, memcpy(dst, src, len) must unify nodes for src and dst.
The reason is that by copying bytes from src, we might copy pointers in
src to dst. However, since the algorithm is flow insensitive, we don't
know whether src has all of its links. So the safe action is to unify.

However, a common usage pattern is to use memcpy on data that does not
have any pointers (i.e., sequence of characters or bytes, etc.). This
commit avoids unifying dst and src when it is reasonably clear that src
has no pointers embedded in it. This should be sound for most common
usages.