Generate always a cell for IntToPtr
This avoids breaking some assumptions during the analysis of other
instructions such as getelementptr in which the base address has
always a cell unless is a global variable. We could avoid generating
the cell if we carefully check all possible users of IntToPtr but I
think it is less error-prone to generate always a cell.