Strip alloca during botom-up pass
If we assume that alloca memory does not escape the function in
which it is allocated, then these allocation sites do not need
to be propagated up the call chain.

Technically, we remove alloca instructions from allocation sites
of every DSNode that is cloned from callee to caller

The infrustructure should also make it possible to enable not
copying alloca durring top-down.

Currently enabled by default. Need an option so that the assumption
can be controlled by command line option