Improve handling on PtrToInt escape (+ clang-format)
Replaced the current hack with complete traversal of def-use
chains of ptrtoint instruction. A ptrtoint instruction escapes
if it flows to Load/Store/Call instruction. Otherwise it does
not escape. There are several common sinks like icmp, br, and sub.

In particular, two pointers are subtracted to compute the size of
allocated memory spanned by them. We assume that the result of the
difference is safe. In examples, it often flows into other functions,
so we need the assumption and cannot rely on def-use chains alone