Improve warning for negative gep offsets
Negative geps are currently not handled. One source of difficulty is
that the order in which Nodes are created is not fixed. Thus we need
to remember at the node that the least most access is negative.

Perhaps gep nodes do not need cells at all. The cell for a gep
can be inferred by computing the barrier for the gep and applying
the appropriate offset from the barrier.

However, in that case there is an issue with handling PHI nodes that
are fed with geps. A PHI node might be fed a negative and a positive
gep.

Current examples only seem to apply negative gep on i8*. It might
indicate a special pattern that can be transformed away.