Support for more inttoptr patterns
Unfortunately, I reformatted DsaLocal with clang-format as well.
But this is good going forward.

A side-effect of this patch is that it might create nodes without allocation
sites. These correspond to dereference  of external memory. External
memory itself might be allocated by external function. But pointers
that are loaded from it end up pointing to potentially unknown memory.