Disable the sorting of callsites.
A CallGraph::CallRecord  (callrecord) is a pair of a call instruction and the call
graph node corresponding to the callee function.

With the CallGraph is produced by the LLVM CallGraph analysis if the
method CallSite::getCallee() when applied on callrecord.first returns
null then callrecord.second->getFunction() must also return null.

For this reason, SortedCallSites assumes that a callsite is irrelevant
to the analysis if CallSite::getCallee() applied on callrecord.first
returns null.

However, with the new call graph resolution each time we know a
potential callee for an indirect call, we keep intact the original
call graph edge but we *add a new call graph edge* to the call graph
node corresponding to the potential callee. (The original edge can be
removed but not always) For that new edge, we will have a callrecord
on which CallSite::getCallee() applied on callrecord.first returns
null but callrecord.second->getFunction() returns a non-null
value. The utility getDsaCallSite deals with this new situation.