Replacing llvm_unreachable with assert(false)
In release mode, llvm_unreachable() are mapped to unreachable marker
and exploited by the optimizer. This creates code that behaves very
strangely (and hard to debug) if an unreachable statement is ever executed
in a real execution.

We tend to use unreachable instead of assertion, but this creates difficult
to debug scenarios when the intended assertion actually fails.