[FIX] dot callgraph.
In Ubuntu, all callgraph nodes were printed as empty strings but on
Mac they were printed correctly.

DOTGraphTraits<GraphType>::getNodeLabel (Support/DOTGraphTraits.h) is
in charge of printing a graph node.  For each graph of type GraphType
that method must be defined, otherwise a default implementation
printing the empty string is used.

The definition of that method for a CallGraph is defined in
lib/Analysis/CallPrinter.cpp.

The solution is to link seadsa with the LLVMAnalysis which contains
CallPrinter.cpp. I don't know how Mac was able to include it.