fix(shadow-mem): incorrect topological traversal
In several places, we assume that llvm::inverse_post_order() can be used
to traverse CFG of a function in topological order.

This is not correct for several reasons:

  (a) a CFG might be cyclic, in which case topo order is not defined

  (b) llvm::inverse_post_order() traverses the inverse CFG. When CFG
      starts from `F.getEntryBlock()`, the root node has no predecessors
      and exploration terminates right after exploring the entry block.

The fix is to traverse the basic blocks in the order they are in the function.
This is, most likely, topological order. In reality, we do not need strict
topological order. It is sufficient that one predecessor has been processed
before the current one.

If this assumption is not correct, we can always add proper topological sort
ignoring back edges at a later time.