feat: Skeleton AliasAnalysis interface for SeaDsa
This is currently the boilerplate skeleton of an analysis. Everything should be
there from the interface perspective, but no analysis is yet hooked up.

The plan is to first hook it up into the pipeline. Since right now it does not
produce any definitive results, this will simply allow us to examine what alias
queries are asked during the run. Then we will hookup the rest.

Note that all LLVM AliasAnalyses are done as `ImmutablePass`. We should follow
the same architecture. In particular, the `CFLAndersAliasAnalysis` and
`CFLSteensAliasAnalysis` are good models.