feat(memssa): API for iterating over memory acceess
`getMemoryAccesses()` functions return a range of memory accesses that are
attached to either a basic block or an instruction.

A basic block might have multiple MemPhi accesses attached to it.

A (complex) instruction might have multiple MemDef or MemUse instructions
attached to it. In particular, functions often read and write multiple memory
regions.