feat(SeaDsaAA): implemented SeaDsaAAResult::alias method.
This is WIP.

There are two different ways to run seaDsaAliasAnalysis: one using
SeaDsaWrapperPass which is an immutable pass and the other by SeaDsaAA
which uses the new AnalysisInfoMixin class. Either way, we need to
pass TLI and AllocWrapInfo analyses to SeaDsaAAResult. I do not know
yet how they can be passed if we use SeaDsaAA. We should get them
somehow from FunctionAnalysisManager.