fix: do not treat external functions as allocators by default
We used to treat any external function returning a pointer as malloc().
This is problematic if we want to use allocation sites to disambiguate
points-to queries.

We would like to say two pointers do not alias if their allocation sites do not
overlap. External functions introduce a gap in def-use chains. We should treat
them as totally unknown. They might return new memory, they might not.

LLVM already provides marker for functions that return new memory and sea-dsa is
already aware of that.