fix(shadow-mem): traverse basic blocks in dfs
To infer shadow-mem annotations on PHI nodes, they must be processed in (at least
DFS) order so that at least one predecessor of a PHI node is processed before
the PHI node itself.

The code assumed that basic blocks are laid out in topological order. But this
is not currently maintained by some of the transformation passes.

Switched to explicit DFS traversal.