feat(shadow-mem): support --horn-sea-dsa-split flag with procedures
This was left as a TODO for a while. Splitting fields was only supported in
intra-procedural mode because fields where not correctly propagated to function.

This commit address this by iterating over all available fields.

The use of --horn-sea-dsa-split assumes that all fields do not overlap.
If fields do overlap, its use might be unsound since it will prevent
interaction between fields that actually do interact.