feat(remove-ptrtoint): handle implicit ptrtoint conversion
The new rewrite handles the following transformation
  %_311 = bitcast i8** @msg_buf to i32*
  %_312 = load i32, i32* %_311, align 4
  %_313 = bitcast %struct.iovec* %_9 to i32*
  store i32 %_312, i32* %_313, align 4

into

  %_311.ptr = bitcast i8** @msg_buf to i8**
  %_312.ptr = load i8*, i8** @_311.ptr, align 4
  %_313.ptr = bitcast %struct.iovec* %_9 to i8**
  store i8* %_312.ptr, i8** %_313.ptr, align 4

The top code writes a pointer as an integer to the beginning of a struct.
The first field of the struct is a pointer, so the effect is setting the
first field to the pointer.

The rewrite code does the same, but uses i8* for the temporary storage instead
of i32