refactor(local): fresh allocation for unsupported constants.
We were returning a null cell which actually can flow to a store or
load and then we get nasty runtime crashes. With this commit, we
instead create a fresh allocation and print a warning message.

Note that I did not add any failing assertion to be consistent with
the code where we are supposed to handle constant struct, arrays,
vectors, etc.