refactor(visitGep): call valueCell before returning null cell
In previous commit, we changed valueCell to return on-the-fly a fresh
cell for unsupported constants to avoid crashing. We modify here
visitGep so that we call valueCell before giving up if there is a not
available a cell for the gep pointer operand. This will avoid crashing
when we have a gep of an unsupported constant expression as pointer
operand.