fix(shadowmem): revert partially commit a393e0e.
This commit tried to fix the problem with issue #102 where we created
a shadow.mem.out annotation also if a region is read-only. However,
the fix in commit a393e0e was too relaxed because created a region
even if the region is not accessed at all. This creates the problem
described in issue #104. The solution that fixes both issues #102 and
#104 is to create a shadow.mem.out if the region is read or modified.