fix(shadowMem): index to relate callsite with function param.
This change only affects the 3rd parameter in shadow.mem.arg.X
(X=ref|mod|new) and shadow.mem.Y (Y=in|out) annotations.

This parameter is an index that allows easily to match a
shadow.mem.arg.X annotation (callsite parameter) at the callsite with
its corresponding shadow.mem.Y (function parameter) in the callee. The
counter used to generate the index was incremented in the callsite at
a different rate that in the callee that caused a mismatch.

For shadow.mem.Y, the counter was incremented even if the node is
neither read or modified.  However, for shadow.mem.arg.X, the counter
was not incremented in that case. The solution is not to increment the
counter if the node is neither read or modified.