fix: support for aliasing between i8* and other types in tea-dsa
Fixed the behaviour of `omnipotent char` flag.

LLVM assumes that many pointers, and, in particular, i8* can
alias. That is, a value written as i8* can be read as another pointer.
This is weaker than the strict aliasing of the C standard. More generally,
each load/store is assigned a tag. Load/store with compatible tags alias.
The tags are partially ordered, so some tags alias with multiple other tags.

This implementation assumes that a pointer to i8* might alias any other type.

Thus, whenver a field with i8* tag exists in a memory object (a node),
it must be unified with all other pointer fields of that object.

A side-effect of the new behavior is that checking whether a field
is present, or adding a new field to an object is more expensive.
A field of a given tag is present if there is a field with a compatible
tag. Adding a field with i8* tag, requires unifying all compatible fields
into one. Current implementation does this on the fly.

A potential alternative is to unify after the fact as a post-processing step.

A good extension of the current implementation is to use TBAA tags instead
of our special case behavior. However, comparison between TBAA tags is
ineffient. This can be solved by caching the tags and their order.

Many little places got patched. Hope that I have gotten them all.

This patch also includes a cleanup of FieldType class.