fix(omni): updating links in Node::addLink when omni is present
The previous code ignored the fact that `unify` can change the current node and
invalidate the `m_links` field that is being traversed.

The new approach uses a more stable interface by adding links via a Cell. This
does appropriate resolution of forwarding if a node gets unified with another
nodes during the operation.