fix(ptrtoint): working around more cases of pointers that escape
Another case of an integer that is used to copy pointer around. Replaced it by
i8*.

If an integer load `LI` is replaced by a pointer load `LP`, then replace all
interesting uses of `LI` by `LP` converted to an integer. This is required for
systems in which pointers are opaque (as in many verification environments).