refactor(seadsa-cg): resolve trivial indirect calls through bitcasts.
When encounter this pattern:

%x = bitcast (void ()* @parse_dir_colors to void (...)*)
call void (...) %x()

The algorithm that constructs the complete callgraph using sea-dsa was
taught to search for the callee through the bitcast but the callgraph
itself was not modified. This is OK for the callgraph construction
algorithm itself but not for other sea-dsa algorithms (e.g., bottom-up
and top-down analyses) which directly traverses the callgraph. As a
result, for the bottom-up analysis if the caller-callee edge was
missing then the bottom-up propagation could happen too early with an
empty graph in the callee if the callee was not yet analyzed. Similar
problem with the top-down analysis.

This problem probably does not occur very often because LLVM
optimizations (if enabled) usually get rid of the code pattern
described above.