fix(callgraph): type-compatibility at callsites.
For checking type-compatibility between callsite and callee we look at
the types returned by getFunctionType(). We need to consider the case
where the callsite or the callee has a vararg type: if either callsite
or callee has a vararg type then the other one must also have a vararg
type.